분리완료
