+++ blockboard/src/main/java/com/board/project/blockboard/mapper/BoardMapper.java
@@ -17,11 +17,6 @@ import java.util.Map;
 @Mapper
 public interface BoardMapper {
     List<BoardDTO> selectBoardsByCompanyID(int companyID);
-    List<PostDTO> selectPostByBoardID(int boardID);
-    PostDTO selectPostByPostID(int postID);
-    int selectCompanyIDByUserID(String userID);
-    String selectCompanyNameByUserID(String userID);
-    String selectUserTypeByUserID(String userID);
     void insertBoard(BoardDTO newBoard);
     BoardDTO selectBoardByBoardName(String newBoardName);
     void updateBoardName(Map<String, Object> boardAttributes);
