+++ blockboard/src/main/java/com/board/project/blockboard/mapper/CompanyMapper.java
@@ -1,0 +1,15 @@
+/**
+ * @author Dongwook Kim <dongwook.kim1211@worksmobile.com>
+ * @file CompanyMapper.java
+ */
+package com.board.project.blockboard.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@Mapper
+public interface CompanyMapper {
+    int selectCompanyIDByUserID(String userID);
+    String selectCompanyNameByUserID(String userID);
+}
