+++ blockboard/src/main/resources/mybatis/mapper/CompanyMapper.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+ * @author Dongwook Kim <dongwook.kim1211@worksmobile.com>
+ * @file CompanyMapper.xml
+-->
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.board.project.blockboard.mapper.CompanyMapper">
+    <select id="selectCompanyNameByUserID" parameterType="string" resultType="string">
+        SELECT company.company_name AS companyName
+        FROM   company company,
+               users users
+        WHERE  company.company_id = users.company_id
+               AND users.user_id = #{userID};
+    </select>
+    <select id="selectCompanyIDByUserID" parameterType="string" resultType="int">
+        SELECT company.company_id AS companyID
+        FROM   company company,
+               users users
+        WHERE  company.company_id = users.company_id
+               AND users.user_id = #{userID};
+    </select>
+</mapper>
