+++ blockboard/src/main/resources/mybatis/mapper/UserMapper.xml
@@ -26,4 +26,12 @@
         FROM Users
         WHERE user_id = #{userID};
     </select>
+
+    <!--userid로 회원 타입 반환-->
+    <!--@author Dongwook Kim <dongwook.kim1211@worksmobile.com>-->
+    <select id="selectUserTypeByUserID" parameterType="string" resultType="string">
+        SELECT user_type AS userType
+        FROM   users
+        WHERE  user_id = #{userID};
+    </select>
 </mapper>
